+++ src/PowerUtils.Validations/Contracts/Rules/PaginationValidationsContract.cs
@@ -4,17 +4,17 @@
     {
         public static IPropertyRule<TSource, string> OrderingDirectionIgnoreCase<TSource>(this IPropertyRule<TSource, string> propertyRule)
         {
-            if (propertyRule.PropertyNull)
+            if(propertyRule.PropertyNull)
             {
                 return propertyRule;
             }
 
-            if (string.IsNullOrEmpty(propertyRule.PropertyValue))
+            if(string.IsNullOrEmpty(propertyRule.PropertyValue))
             {
                 return propertyRule;
             }
 
-            if (propertyRule.PropertyValue.ToLower() != "asc" && propertyRule.PropertyValue.ToLower() != "desc")
+            if(propertyRule.PropertyValue.ToLower() != "asc" && propertyRule.PropertyValue.ToLower() != "desc")
             {
                 propertyRule.AddNotification(ErrorCodes.INVALID);
             }
